refactor: De-duplicate driver code in algorithms
