+++ crates/absolut-macros/src/lib.rs
@@ -1,4 +1,5 @@
 mod algorithms;
+mod driver;
 
 extern crate proc_macro;
 
@@ -6,21 +7,21 @@ use proc_macro::TokenStream;
 
 #[proc_macro_attribute]
 pub fn one_hot(args: TokenStream, tokens: TokenStream) -> TokenStream {
-    crate::algorithms::one_hot::one_hot(args, tokens)
+    driver::driver::<crate::algorithms::one_hot::OneHotBuilder>(args, tokens)
 }
 
 #[proc_macro_attribute]
 pub fn one_cold(args: TokenStream, tokens: TokenStream) -> TokenStream {
-    crate::algorithms::one_cold::one_cold(args, tokens)
+    driver::driver::<crate::algorithms::one_cold::OneColdBuilder>(args, tokens)
 }
 
 #[proc_macro_attribute]
 pub fn composite(args: TokenStream, tokens: TokenStream) -> TokenStream {
-    crate::algorithms::composite::composite(args, tokens)
+    driver::driver::<crate::algorithms::composite::CompositeBuilder>(args, tokens)
 }
 
 #[cfg(feature = "general")]
 #[proc_macro_attribute]
 pub fn general(args: TokenStream, tokens: TokenStream) -> TokenStream {
-    crate::algorithms::general::general(args, tokens)
+    driver::driver::<crate::algorithms::general::GeneralBuilder>(args, tokens)
 }
